requirements.txt
# Python 3.9.13
accelerate==0.18.0
albumentations==0.5.2
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.6.0
executing==1.2.0
filelock==3.11.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
imgaug==0.4.0
importlib-metadata==6.4.1
importlib-resources==5.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
lion-pytorch==0.0.7
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
munch==2.5.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==1.5.1
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
pywin32==305.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.1.3
smmap==5.0.0
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.13
# torch==2.0.0+cu117
# torchvision==0.15.0+cu117
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.13.11
wcwidth==0.2.6
wheel==0.38.4
wincertstore==0.2
zipp==3.15.0